+++ platform/OHOS/sweetline/src/main/cpp/sweetline_bindings.cpp
@@ -577,6 +577,25 @@ static napi_value DocumentAnalyzer_AnalyzeChangesInLineRange(napi_env env, napi_
   return ConvertDocumentHighlightSliceAsIntArray(env, analyzer->getHighlightConfig(), slice);
 }
 
+/// Get highlight slice from the current cached result
+/// args: [handle, visibleStartLine, visibleLineCount]
+static napi_value DocumentAnalyzer_GetHighlightSlice(napi_env env, napi_callback_info info) {
+  size_t argc = 3;
+  napi_value args[3] = {nullptr};
+  napi_get_cb_info(env, info, &argc, args, nullptr, nullptr);
+
+  SharedPtr<DocumentAnalyzer> analyzer = getNapiCPtrHolderValue<DocumentAnalyzer>(env, args[0]);
+  if (analyzer == nullptr) {
+    return getNapiUndefined(env);
+  }
+  int32_t visible_start_line = 0, visible_line_count = 0;
+  napi_get_value_int32(env, args[1], &visible_start_line);
+  napi_get_value_int32(env, args[2], &visible_line_count);
+  LineRange visible_range = {static_cast<size_t>(visible_start_line), static_cast<size_t>(visible_line_count)};
+  SharedPtr<DocumentHighlightSlice> slice = analyzer->getHighlightSlice(visible_range);
+  return ConvertDocumentHighlightSliceAsIntArray(env, analyzer->getHighlightConfig(), slice);
+}
+
 /// Perform indent guide analysis on the managed document
 static napi_value DocumentAnalyzer_AnalyzeIndentGuides(napi_env env, napi_callback_info info) {
   size_t argc = 1;
@@ -856,6 +875,7 @@ static napi_value Init(napi_env env, napi_value exports) {
     {"DocumentAnalyzer_AnalyzeChanges", nullptr, DocumentAnalyzer_AnalyzeChanges, nullptr, nullptr, nullptr, napi_default, nullptr},
     {"DocumentAnalyzer_AnalyzeChanges2", nullptr, DocumentAnalyzer_AnalyzeChanges2, nullptr, nullptr, nullptr, napi_default, nullptr},
     {"DocumentAnalyzer_AnalyzeChangesInLineRange", nullptr, DocumentAnalyzer_AnalyzeChangesInLineRange, nullptr, nullptr, nullptr, napi_default, nullptr},
+    {"DocumentAnalyzer_GetHighlightSlice", nullptr, DocumentAnalyzer_GetHighlightSlice, nullptr, nullptr, nullptr, napi_default, nullptr},
     {"DocumentAnalyzer_AnalyzeIndentGuides", nullptr, DocumentAnalyzer_AnalyzeIndentGuides, nullptr, nullptr, nullptr, napi_default, nullptr},
     {"DocumentAnalyzer_GetDocument", nullptr, DocumentAnalyzer_GetDocument, nullptr, nullptr, nullptr, napi_default, nullptr},
     // HighlightEngine