+++ platform/OHOS/sweetline/src/main/cpp/types/libsweetline/Index.d.ts
@@ -26,6 +26,7 @@ export const DocumentAnalyzer_Analyze: (handle: number) => Int32Array;
 export const DocumentAnalyzer_AnalyzeChanges: (handle: number, startLine: number, startColumn: number, endLine: number, endColumn: number, newText: string) => Int32Array;
 export const DocumentAnalyzer_AnalyzeChanges2: (handle: number, startIndex: number, endIndex: number, newText: string) => Int32Array;
 export const DocumentAnalyzer_AnalyzeChangesInLineRange: (handle: number, startLine: number, startColumn: number, endLine: number, endColumn: number, newText: string, visibleStartLine: number, visibleLineCount: number) => Int32Array;
+export const DocumentAnalyzer_GetHighlightSlice: (handle: number, visibleStartLine: number, visibleLineCount: number) => Int32Array;
 export const DocumentAnalyzer_AnalyzeIndentGuides: (handle: number) => Int32Array;
 export const DocumentAnalyzer_GetDocument: (handle: number) => number;